move entitySetting-utils to utils directory
