+++ opencti-platform/opencti-graphql/src/modules/entitySetting/entitySetting-resolvers.ts
@@ -13,7 +13,7 @@ import {
 import type { Resolvers } from '../../generated/graphql';
 import { BUS_TOPICS } from '../../config/conf';
 import { ENTITY_TYPE_ENTITY_SETTING } from './entitySetting-types';
-import { getAvailableSettings } from './entitySetting-utils';
+import { getAvailableSettings } from '../../utils/entitySetting-utils';
 import { subscribeToInstanceEvents } from '../../graphql/subscriptionWrapper';
 import { getRequestAccessConfiguration } from '../requestAccess/requestAccess-domain';
