+++ opencti-platform/opencti-graphql/src/modules/internal/jsonMapper/jsonMapper-utils.ts
@@ -28,7 +28,7 @@ import { internalFindByIds } from '../../../database/middleware-loader';
 import type { BasicStoreEntity } from '../../../types/store';
 import { extractRepresentative } from '../../../database/entity-representative';
 import { FunctionalError } from '../../../config/errors';
-import { fillDefaultValues, getEntitySettingFromCache } from '../../entitySetting/entitySetting-utils';
+import { fillDefaultValues, getEntitySettingFromCache } from '../../../utils/entitySetting-utils';
 import { schemaAttributesDefinition } from '../../../schema/schema-attributes';
 import { isEmptyField, isNotEmptyField } from '../../../database/utils';
 import { representationLabel } from '../mapper-utils';
